+++ src/mpd/proto_client.rs
@@ -56,6 +56,7 @@ impl<'cmd, 'client, C: SocketClient> ProtoClient<'cmd, 'client, C> {
         trace!(command = self.command; "Executing command");
         if let Err(e) = self.client.write([command, "\n"].concat().as_bytes()) {
             if e.kind() == std::io::ErrorKind::BrokenPipe {
+                log::error!(err:? = e; "Got broken pipe from mpd");
                 self.client.reconnect()?;
                 self.client.write([command, "\n"].concat().as_bytes())?;
                 Ok(self)
@@ -282,12 +283,21 @@ impl<'cmd, 'client, C: SocketClient> ProtoClient<'cmd, 'client, C> {
 
         let bytes_read = match read.read_line(&mut line) {
             Ok(v) => Ok(v),
-            Err(e) if e.kind() == std::io::ErrorKind::BrokenPipe => Err(MpdError::ClientClosed),
-            _ => Err(MpdError::ClientClosed),
+            Err(e) if e.kind() == std::io::ErrorKind::BrokenPipe => {
+                log::error!(err:? = e; "Got broken pipe from mpd");
+                Err(MpdError::ClientClosed)
+            }
+            Err(e) => {
+                log::error!(err:? = e; "Encountered unexpected error whe reading a response  line from MPD");
+                Err(e.into())
+            }
         }?;
 
         if bytes_read == 0 {
-            return Err(MpdError::ClientClosed);
+            log::error!("Got an empty line in MPD's response");
+            return Err(MpdError::ValueExpected(
+                "Expected value when reading MPD's response but the stream reached EOF".to_string(),
+            ));
         }
 
         if line.starts_with("OK") || line.starts_with("list_OK") {